restructure nav component and update part of styling
